refactor: добавление аннотации изменение списка на мапу

+++ main/src/main/java/ru/practicum/events/dto/EventFullDto.java
@@ -26,5 +26,5 @@ public class EventFullDto {
     private Boolean requestModeration;
     private String state;
     private String title;
-    private int views;
+    private Long views;
 }
