+++ main/src/main/java/ru/practicum/events/dto/ViewStatisticDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class ViewStatisticDto {
     private String app;
     private String uri;
-    private int hits;
+    private long hits;
 }
